// packages/web/src/types/enums.ts
export enum Games {
  AssassinCreditValhalla = 'assassin_credit_valhalla',
  CodWarzone = 'cod_warzone',
  GenshinImpact = 'genshin_impact',
  Pubg = 'pubg',
  ZeldaBow = 'zelda_bow',
  GhostOfTsushima = 'ghost_of_tsushima',
  HollowKnight = 'hollow_knight',
}

export enum OSS {
  QINIU = 'http://static.shuotu.fun',
  QINGYUN = 'https://tushuo.sh1a.qingstor.com',
}